requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
google-generativeai==0.3.2
flask==2.3.2
gunicorn==21.2.0